get_time: testing version

+++ src/drivers/get_time.h
@@ -1,12 +1,12 @@
 #ifndef GET_TIME_H
 #define GET_TIME_H
 
-uint32_t start_time_ms();
-uint32_t stop_time_ms();
+uint16_t start_time_ms();
+uint16_t stop_time_ms();
 
-float convert_cycles_to_time(unsigned int cycles);
+uint32_t convert_cycles_to_time(uint32_t cycles);
 void start_time();
-float stop_time();
-unsigned int getCycles();
+uint32_t stop_time();
+uint32_t getCycles();
 
 #endif /* GET_TIME_H */
